Make hic rendering
