+++ packages/hic/src/HicAdapter/declare.d.ts
@@ -1,0 +1,1 @@
+declare module 'hic-straw'
