+++ packages/hic/src/HicRenderer/HicRenderer.ts
@@ -1,0 +1,88 @@
+import { AnyConfigurationModel } from '@gmod/jbrowse-core/configuration/configurationSchema'
+import ServerSideRendererType from '@gmod/jbrowse-core/pluggableElementTypes/renderers/ServerSideRendererType'
+import { Feature } from '@gmod/jbrowse-core/util/simpleFeature'
+import { Region } from '@gmod/jbrowse-core/util/types'
+import {
+  createCanvas,
+  createImageBitmap,
+} from '@gmod/jbrowse-core/util/offscreenCanvasPonyfill'
+import React from 'react'
+import { readConfObject } from '@gmod/jbrowse-core/configuration'
+
+export interface PileupRenderProps {
+  features: Map<string, Feature>
+  config: AnyConfigurationModel
+  regions: Region[]
+  bpPerPx: number
+  height: number
+  width: number
+  highResolutionScaling: number
+  sortObject: {
+    position: number
+    by: string
+  }
+}
+
+export default class HicRenderer extends ServerSideRendererType {
+  async makeImageData(props: PileupRenderProps) {
+    const {
+      features,
+      config,
+      regions,
+      bpPerPx,
+      highResolutionScaling = 1,
+    } = props
+    const [region] = regions
+    console.log(features)
+
+    const width = (region.end - region.start) / bpPerPx
+    const height = readConfObject(config, 'maxHeight')
+    if (!(width > 0) || !(height > 0))
+      return { height: 0, width: 0, maxHeightReached: false }
+
+    const canvas = createCanvas(
+      Math.ceil(width * highResolutionScaling),
+      height * highResolutionScaling,
+    )
+    const ctx = canvas.getContext('2d')
+    ctx.scale(highResolutionScaling, highResolutionScaling)
+    ctx.font = 'bold 10px Courier New,monospace'
+    const charSize = ctx.measureText('A')
+    charSize.height = 7
+
+    const imageData = await createImageBitmap(canvas)
+    return {
+      imageData,
+      height,
+      width,
+    }
+  }
+
+  async render(renderProps: PileupRenderProps) {
+    const {
+      height,
+      width,
+      imageData,
+      maxHeightReached,
+    } = await this.makeImageData(renderProps)
+    const element = React.createElement(
+      this.ReactComponent,
+      {
+        ...renderProps,
+        region: renderProps.regions[0],
+        height,
+        width,
+        imageData,
+      },
+      null,
+    )
+
+    return {
+      element,
+      imageData,
+      height,
+      width,
+      maxHeightReached,
+    }
+  }
+}
