+++ packages/hic/src/HicRenderer/components/HicRendering.test.js
@@ -1,12 +1,12 @@
 import PrecomputedLayout from '@gmod/jbrowse-core/util/layouts/PrecomputedLayout'
 import React from 'react'
 import { render } from '@testing-library/react'
-import PileupRendering from './PileupRendering'
+import HicRendering from './HicRendering'
 
 // these tests do very little, let's try to expand them at some point
 test('one', () => {
   const { container } = render(
-    <PileupRendering
+    <HicRendering
       width={500}
       height={500}
       regions={[{ refName: 'zonk', start: 1, end: 3 }]}
