+++ packages/hic/src/HicRenderer/index.ts
@@ -1,3 +1,3 @@
-export { default as ReactComponent } from './components/PileupRendering'
+export { default as ReactComponent } from './components/HicRendering'
 export { default as configSchema } from './configSchema'
-export { default } from './PileupRenderer'
+export { default } from './HicRenderer'
